// src/app/model/addbusinessdetail.ts
export class BusineesDetail {
 ID: number;
 BusinessName: string;
 BusinessAddress1: string;
 BusinessAddress2: string;
 CityID:  number;
 ContryID:  number;
 PostCode: string;
 Phone: string;
 Email: string;
 Website: string;
 EmailVerified: boolean;
 PhoneVerified: boolean;
 StateID: number;
 AdditionalDetails: string;
 UserID: number;
}